Added Polygon pack foundation + Polygonal tag
Added a new booster pack for Polygon Monsters, a lil branch of Polygonal-based creatures. Very WIP, but the pack itself is in.

Next up are the cards themselves and then after trying to make the pack act more like the buffoon or standard packs where you choose rather than use.