2 new Polygon vouchers and Polygon shader call fix
Added Bestiary and Grimoire vouchers, which both decrease the amount of hands it takes to play Polygon cards.

If needed rounds < 0, then they are free to be used immediately.

Also fixed the shader call that applies the smoke shader onto the polygon set. It now applies it during the event that juices the card, instead of immediately upon evaluation.